Convert deltatime to float before using

+++ Components/Component.cs
@@ -51,7 +51,7 @@ namespace KEngine.Components {
 
         }
 
-        public virtual void Update(double deltaTime) {
+        public virtual void Update(float deltaTime) {
 
         }
